separate git and github functionality
