+++ abcclassroom/github.py
@@ -3,21 +3,16 @@ abc-classroom.github
 ====================
 """
 
-import os
-import logging
-import random
-import string
-import subprocess
-import sys
+# Methods for setting up and accessing the GitHub API. See
+# git.py for command line git operations.
+
 import requests
 
 import os.path as op
-
 from ruamel.yaml import YAML
-
 import github3 as gh3
 
-from .utils import input_editor, get_request
+from .utils import get_request
 
 
 def get_github_auth():
@@ -108,65 +103,6 @@ def get_access_token():
     return access_token
 
 
-def check_git_ssh():
-    """Tests that ssh access to GitHub is set up correctly on the users
-    computer.
-
-    Throws a RuntimeError if setup is not working.
-    """
-    cmd = ["ssh", "-T", "git@github.com"]
-    try:
-        subprocess.run(
-            cmd,
-            check=True,
-            universal_newlines=True,
-            stdout=subprocess.PIPE,
-            stderr=subprocess.PIPE,
-        )
-    except subprocess.CalledProcessError as e:
-        # We ALWAYS will get here, because that subprocess call returns
-        # a non-zero exit code whether ssh access is set up correctly or
-        # not. Must check output.
-        subprocess_out = e.stderr
-        if not subprocess_out:
-            subprocess_out = e.stdout
-        if subprocess_out.startswith("Hi"):
-            # if everything set up correctly, the ssh test returns output
-            # starting with 'Hi username!'
-            pass
-        elif subprocess_out.startswith("Warning: Permanently"):
-            # if the user has set up ssh keys, but hasn't logged in yet,
-            # they will need to verify the RSA fingerprint
-            # If they do so, the message is 'Warning: Permanently
-            # added 'github.com' (RSA) to the list of known hosts.'
-            print(subprocess_out)
-            pass
-        else:
-            # possible reasons to get here include 1. no ssh key set up;
-            # 2. ssh key has incorrect permissions; 3. remote host
-            # identification changed. We print the error message for
-            # the user and point them to github documentation for help.
-            print("Encountered this error checking ssh access to git:")
-            print(subprocess_out)
-            docURL = "https://docs.github.com/en/github/authenticating-to-github/connecting-to-github-with-ssh"  # noqa
-            print(
-                """Your ssh access to github is not set up correctly; see
-            {}.""".format(
-                    docURL
-                )
-            )
-            raise RuntimeError(subprocess_out)
-    except FileNotFoundError as e:
-        # we get here if ssh is not installed. The error message is
-        # [Errno 2] No such file or directory: 'ssh'
-        print(e)
-        if "No such file or directory: 'ssh'" in str(e):
-            print(
-                """Did not find `ssh` command. Make sure that open-ssh
-                is installed on your operating system."""
-            )
-
-
 def _get_authenticated_user(token):
     """Test the validity of an access token.
 
@@ -248,27 +184,6 @@ def _poll_for_status(client_id, device_code):
     return access_token
 
 
-def _call_git(*args, directory=None):
-    cmd = ["git"]
-    cmd.extend(args)
-    try:
-        ret = subprocess.run(
-            cmd,
-            cwd=directory,
-            check=True,
-            universal_newlines=True,
-            stdout=subprocess.PIPE,
-            stderr=subprocess.PIPE,
-        )
-    except subprocess.CalledProcessError as e:
-        err = e.stderr
-        if not err:
-            err = e.stdout
-        raise RuntimeError(err) from e
-
-    return ret
-
-
 def remote_repo_exists(org, repository, token=None):
     """Check if the remote repository exists for the organization."""
 
@@ -280,23 +195,6 @@ def remote_repo_exists(org, repository, token=None):
         return False
 
     return True
-
-
-def clone_repo(organization, repo, dest_dir):
-    """Clone `repository` from `org` into a sub-directory in `directory`.
-
-    Raises RuntimeError if ssh keys not set up correctly, or if git clone
-    fails for other reasons.
-    """
-
-    try:
-        # first, check that local git set up with ssh keys for github
-        check_git_ssh()
-        url = "git@github.com:{}/{}.git".format(organization, repo)
-        print("cloning:", url)
-        _call_git("-C", dest_dir, "clone", url)
-    except RuntimeError as e:
-        raise e
 
 
 def create_repo(org, repository, token):
@@ -318,242 +216,102 @@ def create_repo(org, repository, token):
         )
 
 
-def add_remote(directory, organization, remote_repo):
-    remote_url = "git@github.com:{}/{}.git".format(organization, remote_repo)
-    _call_git("remote", "add", "origin", remote_url, directory=directory)
-
-
-def repo_changed(directory):
-    """Determine if the Git repository in directory is dirty"""
-    ret = _call_git("status", "--porcelain", directory=directory)
-    return bool(ret.stdout)
-
-
-def new_branch(directory, name=None):
-    """Create a new git branch in directory"""
-    if name is None:
-        postfix = "".join(
-            [random.choice(string.ascii_letters) for n in range(4)]
-        )
-        name = "new-material-{}".format(postfix)
-
-    _call_git("checkout", "-b", name, directory=directory)
-
-    return name
-
-
-def get_commit_message():
-    default_message = """
-    # Please enter the commit message for your changes. Lines starting
-    # with '#' will be ignored, and an empty message aborts the commit.
-    # This message will be used as commit and Pull Request message."""
-    message = input_editor(default_message)
-    message = "\n".join(
-        [
-            line
-            for line in message.split("\n")
-            if not line.strip().startswith("#")
-        ]
-    )
-    return message
-
-
-def commit_all_changes(directory, msg=None):
-    """Run git add, git commit on a given directory. Checks git status
-    first and does nothing if no changes.
-    """
-    if msg is None:
-        raise ValueError("Commit message can not be empty.")
-    if repo_changed(directory):
-        _call_git("add", "*", directory=directory)
-        _call_git("commit", "-a", "-m", msg, directory=directory)
-    else:
-        print("No changes in repository {}; doing nothing".format(directory))
-
-
-def init_and_commit(directory, custom_message=False):
-    """Run git init, git add, git commit on given directory. Checks git status
-    first and does nothing if no changes.
-    """
-    # local git things - initialize, add, commit
-    # note that running git init on an existing repo is safe, so no need
-    # to check anything first
-    git_init(directory)
-    _master_branch_to_main(directory)
-    if repo_changed(directory):
-        message = "Initial commit"
-        if custom_message:
-            message = get_commit_message()
-            if not message:
-                print("Empty commit message, exiting.")
-                sys.exit(1)  # sys is undefined - ask karen about this
-        commit_all_changes(directory, message)
-    else:
-        print("No changes to local repository.")
-
-
-def _master_branch_to_main(dir):
-    """Change the name of the master branch to main
-
-    Changes the name of the master branch to main for the repo in the
-    given directory. Since we create the repo on github first, which now sets
-    the default branch to 'main', we need the local repo to match
-    in order to be able to push without errors later.
-    """
-
-    try:
-        # first, verify if the  master branch exists (this is only true
-        # if there are commits on the master branch)
-        _call_git(
-            "show-ref",
-            "--quiet",
-            "--verify",
-            "refs/heads/master",
-            directory=dir,
-        )
-    except RuntimeError:
-        # master branch verification fails, so we check for main and create
-        # it if it does not exist
-        try:
-            _call_git(
-                "show-ref",
-                "--quiet",
-                "--verify",
-                "refs/heads/main",
-                directory=dir,
-            )
-        except RuntimeError:
-            # no main branch, create one
-            _call_git("checkout", "-b", "main", directory=dir)
-    else:
-        # rename branch
-        print("master exists, renaming")
-        _call_git("branch", "-m", "master", "main", directory=dir)
-
-
-def push_to_github(directory, branch="main"):
-    """Push `branch` of the repository in `directory` back to GitHub"""
-    try:
-        # first, check that local git set up with ssh keys for github
-        check_git_ssh()
-        _call_git(
-            "push", "--set-upstream", "origin", branch, directory=directory
-        )
-    except RuntimeError as e:
-        raise e
-
-
-def pull_from_github(directory, branch="main"):
-    """Pull `branch` of local repo in `directory` from GitHub"""
-    try:
-        # first, check that local git set up with ssh keys for github
-        check_git_ssh()
-        _call_git("pull", "origin", branch, directory=directory)
-    except RuntimeError as e:
-        raise e
-
-
-def git_init(directory, defaultbranch="main"):
-    """Initialize git repository"""
-    _call_git("init", directory=directory)
-
-
 ###################################################
 # Methods below are from before the re-factoring.
 # Retaining for reference, but with no guarantee
 # about correct function.
 
 
-def check_student_repo_exists(org, course, student, token=None):
-    """Check if the student has a repository for the course.
-
-    It happens that students delete their repository or do not accept the
-    invitation to the course. In either case they will not have a repository
-    yet.
-    """
-    # temporarily change log level of github3.py as it prints weird messages
-    # XXX could be done more nicely with a context manager maybe
-    gh3_log = logging.getLogger("github3")
-    old_level = gh3_log.level
-    gh3_log.setLevel("ERROR")
-
-    try:
-        g = gh3.login(token=token)
-        repository = "{}-{}".format(course, student)
-        g.repository(org, repository)
-
-    except Exception as e:
-        raise e
-
-    finally:
-        gh3_log.setLevel(old_level)
-
-
-def close_existing_pullrequests(
-    org, repository, branch_base="new-material-", token=None
-):
-    """Close all oustanding course material update Pull Requests
-
-    If there are any PRs open in a student's repository that originate from
-    a branch starting with `branch_base` as name and created by the user
-    we are logged in we close them.
-    """
-    g = gh3.login(token=token)
-    me = g.me()
-    repo = g.repository(org, repository)
-    for pr in repo.pull_requests(state="open"):
-        origin = pr.head.label
-        origin_repo, origin_branch = origin.split(":")
-        if origin_branch.startswith(branch_base) and pr.user == me:
-            pr.create_comment(
-                "Closed in favor of a new Pull Request to "
-                "bring you up-to-date."
-            )
-            pr.close()
-
-
-def create_pr(org, repository, branch, message, token):
-    """Create a Pull Request with changes from branch"""
-    msg_parts = message.split("\n\n")
-    if len(msg_parts) == 1:
-        title = msg = msg_parts[0]
-    else:
-        title = msg_parts[0]
-        msg = "\n\n".join(msg_parts[1:])
-
-    g = gh3.login(token=token)
-    repo = g.repository(org, repository)
-    repo.create_pull(title, "master", branch, msg)
-
-
-def fetch_student(org, course, student, directory, token=None):
-    """Fetch course repository for `student` from `org`
-
-    The repository will be cloned into a sub-directory in `directory`.
-
-    Returns the directory in which to find the students work.
-    """
-    # use ssh if there is no token
-    if token is None:
-        fetch_command = [
-            "git",
-            "clone",
-            "git@github.com:{}/{}-{}.git".format(org, course, student),
-        ]
-    else:
-        fetch_command = [
-            "git",
-            "clone",
-            "https://{}@github.com/{}/{}-{}.git".format(
-                token, org, course, student
-            ),
-        ]
-    subprocess.run(
-        fetch_command,
-        cwd=directory,
-        check=True,
-        stdout=subprocess.PIPE,
-        stderr=subprocess.PIPE,
-    )
-
-    return os.path.join(directory, "{}-{}".format(course, student))
+# def check_student_repo_exists(org, course, student, token=None):
+#     """Check if the student has a repository for the course.
+#
+#     It happens that students delete their repository or do not accept the
+#     invitation to the course. In either case they will not have a repository
+#     yet.
+#     """
+#     # temporarily change log level of github3.py as it prints weird messages
+#     # XXX could be done more nicely with a context manager maybe
+#     gh3_log = logging.getLogger("github3")
+#     old_level = gh3_log.level
+#     gh3_log.setLevel("ERROR")
+#
+#     try:
+#         g = gh3.login(token=token)
+#         repository = "{}-{}".format(course, student)
+#         g.repository(org, repository)
+#
+#     except Exception as e:
+#         raise e
+#
+#     finally:
+#         gh3_log.setLevel(old_level)
+#
+#
+# def close_existing_pullrequests(
+#     org, repository, branch_base="new-material-", token=None
+# ):
+#     """Close all oustanding course material update Pull Requests
+#
+#     If there are any PRs open in a student's repository that originate from
+#     a branch starting with `branch_base` as name and created by the user
+#     we are logged in we close them.
+#     """
+#     g = gh3.login(token=token)
+#     me = g.me()
+#     repo = g.repository(org, repository)
+#     for pr in repo.pull_requests(state="open"):
+#         origin = pr.head.label
+#         origin_repo, origin_branch = origin.split(":")
+#         if origin_branch.startswith(branch_base) and pr.user == me:
+#             pr.create_comment(
+#                 "Closed in favor of a new Pull Request to "
+#                 "bring you up-to-date."
+#             )
+#             pr.close()
+#
+#
+# def create_pr(org, repository, branch, message, token):
+#     """Create a Pull Request with changes from branch"""
+#     msg_parts = message.split("\n\n")
+#     if len(msg_parts) == 1:
+#         title = msg = msg_parts[0]
+#     else:
+#         title = msg_parts[0]
+#         msg = "\n\n".join(msg_parts[1:])
+#
+#     g = gh3.login(token=token)
+#     repo = g.repository(org, repository)
+#     repo.create_pull(title, "master", branch, msg)
+#
+#
+# def fetch_student(org, course, student, directory, token=None):
+#     """Fetch course repository for `student` from `org`
+#
+#     The repository will be cloned into a sub-directory in `directory`.
+#
+#     Returns the directory in which to find the students work.
+#     """
+#     # use ssh if there is no token
+#     if token is None:
+#         fetch_command = [
+#             "git",
+#             "clone",
+#             "git@github.com:{}/{}-{}.git".format(org, course, student),
+#         ]
+#     else:
+#         fetch_command = [
+#             "git",
+#             "clone",
+#             "https://{}@github.com/{}/{}-{}.git".format(
+#                 token, org, course, student
+#             ),
+#         ]
+#     subprocess.run(
+#         fetch_command,
+#         cwd=directory,
+#         check=True,
+#         stdout=subprocess.PIPE,
+#         stderr=subprocess.PIPE,
+#     )
+#
+#     return os.path.join(directory, "{}-{}".format(course, student))
